Added sample tests
