+++ src/main/resources/templates/register.html
@@ -8,17 +8,17 @@
 <body>
 <h2>Register</h2>
 <form id="registerform" th:action="@{/createUser}">
-    First Name: <input type="text" id="registerFirstName"> <br>
-    Last Name: <input type="text" id="registerLastName"> <br>
-    Email: <input type="text" id="registerEmail"> <br>
-    Username: <input type="text" id="registerUsername"> <br>
-    Password: <input type="text" id="registerPassword"> <br>
+    First Name: <input type="text" id="registerFirstName" th:value="*{firstName}"> <br>
+    Last Name: <input type="text" id="registerLastName" th:value="*{lastName}"> <br>
+    Email: <input type="text" id="registerEmail" th:value="*{email}"> <br>
+    Username: <input type="text" id="registerUsername" th:value="*{username}"> <br>
+    Password: <input type="text" id="registerPassword" th:value="*{password}"> <br>
     Type:
     <select>
         <option value="editor">Editor</option>
         <option value="reviewer">Reviewer</option>
     </select> <br>
-    <button type="button" id="registerBtn">Register</button>
+    <button type="button" id="registerBtn" value="home">Register</button>
 </form>
 </body>
 <!--<script src="script.js" type="text/javascript"></script>-->
